+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/test/ui/details/TestCertificateDetailsFragment.kt
@@ -117,7 +117,7 @@ class TestCertificateDetailsFragment : Fragment(R.layout.fragment_test_certifica
             )
             is TestCertificateDetailsNavigation.ValidationStart -> doNavigate(
                 TestCertificateDetailsFragmentDirections
-                    .actionTestCertificateDetailsFragmentToValidationStartFragment()
+                    .actionTestCertificateDetailsFragmentToValidationStartFragment(event.containerId)
             )
         }
     }
