+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/test/ui/details/TestCertificateDetailsNavigation.kt
@@ -1,7 +1,9 @@
 package de.rki.coronawarnapp.covidcertificate.test.ui.details
 
+import de.rki.coronawarnapp.covidcertificate.common.repository.CertificateContainerId
+
 sealed class TestCertificateDetailsNavigation {
     object Back : TestCertificateDetailsNavigation()
     data class FullQrCode(val qrCodeText: String) : TestCertificateDetailsNavigation()
-    object ValidationStart : TestCertificateDetailsNavigation()
+    data class ValidationStart(val containerId: CertificateContainerId) : TestCertificateDetailsNavigation()
 }
