+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/covidcertificate/test/ui/details/TestCertificateDetailsViewModel.kt
@@ -39,7 +39,7 @@ class TestCertificateDetailsViewModel @AssistedInject constructor(
 
     fun openFullScreen() = qrCodeText?.let { events.postValue(TestCertificateDetailsNavigation.FullQrCode(it)) }
 
-    fun onValidationStart() = events.postValue(TestCertificateDetailsNavigation.ValidationStart)
+    fun onValidationStart() = events.postValue(TestCertificateDetailsNavigation.ValidationStart(containerId))
 
     fun onDeleteTestCertificateConfirmed() = launch {
         Timber.d("Removing Test Certificate=$containerId")
